# server/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API Computación_UTM",
        "description": "<html><body>Esta es la documentación de la API para un sistema de control de los Alumnos, Profesores y Salones de la Carrera de Computación en la UTM. <p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los  servicios proporcionados.</p> <br> <p> <b>Autenticación y token de API</b></p><p>El administrador del sistema de la Carrera de Computación de la UTM proporciona el api-token una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/Computacion/Alumnos/{matricula}": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Muestra la información de un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de un los Alumno que pertenecen a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuáles son los Alumnos?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "matricula": 1,
                                    "nombre": "Arturo Aquino",
                                    "edad": 21,
                                    "sexo": "masculino",
                                    "Grado": 5
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Alumnos"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "matricula",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Profesores/{id}": {
            "get": {
                "tags": [
                    "Profesores"
                ],
                "summary": "Muestra la información de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve la información detallada de un Profesor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál es el Profesor?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id": 1,
                                    "nombre": "Eduardo Soto",
                                    "edad": 50,
                                    "sexo": "masculino",
                                    "materias_impartidas": 2,
                                    "correo": "eduardosoto@gs.utm.mx",
                                    "tipo":2
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Profesores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Profesores/Horarios/{id}": {
            "get": {
                "tags": [
                    "Horarios Profesores"
                ],
                "summary": "Muestra la información de los Horarios de un Profesor de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve los Horarios detallados de un Profesor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál Horario de Profesor se agrega?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id": 6,
                                    "nombre": "Moises Ramírez",
                                    "materias impartidas": "Computo Reconfigurable",
                                    "grado": 4,
                                    "num_salon": "Aula 7",
                                    "hora":"9:00am-10:00am"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Horarios_Profesores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Profesor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Alumnos/Horarios/{matricula}": {
            "get": {
                "tags": [
                    "Horarios Alumnos"
                ],
                "summary": "Muestra la información de los Horarios de un Alumno de la Carrera de Computación de la UTM dado su matrícula.",
                "description": "<p>Servicio que devuelve los Horarios detallados de un Alumno perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p><p><em><strong>¿Cuál Horario del Alumno se agrega?</strong></em></p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "matricula": 1,
                                    "nombre": "Arturo Aquino",
                                    "materia": "Computo Reconfigurable",
                                    "hora":"9:00am-10:00am",
                                    "salon":"Aula 8"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Horarios_Alumnos"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Profesores/Tutorados/{id}": {
            "get": {
                "tags": [
                    "Tutorados"
                ],
                "summary": "Muestra la información de los Tutorados de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de un Tutorado perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "tutor": "Moises Ramírez",
                                    "id_alu": 1,
                                    "nombre_alumno": "Arturo Aquino"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutorados"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Tutorias/{id_al}": {
            "get": {
                "tags": [
                    "Tutoria"
                ],
                "summary": "Muestra la información de una Tutoría a un Alumno de la Carrera de Computación de la UTM dada su matrícula.",
                "description": "<p>Servicio que devuelve la información detallada de una Tutoría a un Alumno perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "tutor": "Moises Ramírez",
                                    "id_alu": 1,
                                    "nombre_alumno": "Arturo Aquino"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutoria"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere matrícula del Alumno",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
        "/api/Computacion/Alumnos/Tutores/{id}": {
            "get": {
                "tags": [
                    "Tutores"
                ],
                "summary": "Muestra la información de los Tutores de la Carrera de Computación de la UTM dado su id.",
                "description": "<p>Servicio que devuelve la información detallada de un Tutor perteneciente a la Carrera de Computación de la UTM </p><p>&nbsp;</p>",
                "operationId": "listOne",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "id_prof": 1,
                                    "nombre_profesor": "Erik Ramos",
                                    "horario_asesoria": "5:00pm-6:00pm",
                                    "atutorados": "Omar Cruz"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Tutores"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "requiere id del Tutor",
                        "type": "number",
                        "default": "1"
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Profesores": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string",
                    "description": "Describe el nombre de los Profesores"
                },
                "edad": {
                    "type": "number",
                    "description": "Describe la edad del Profesor"
                },
                "materias_impartidas": {
                    "type": "string",
                    "description": "Describe la cantidad materias que imparte el Profesor"
                },
                "sexo": {
                    "type": "string",
                    "description": "Describe el seco del profesor"
                },
                "password": {
                    "type": "string",
                    "description": "Contraseña de acceso al sistema"
                },
                "correo": {
                    "type": "string",
                    "description": "Correo para acceder al sistema (único del Profesor)"
                },
                "tipo":{
                    "type": "number",
                    "description": "Describe el puesto de un Profesor"
                }
            }
        },
        "Alumnos": {
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string",
                    "description": "Describe el nombre de los Alumnos"
                },
                "edad": {
                    "type": "number",
                    "description": "Describe la edad de los Alumnos"
                },
                "sexo": {
                    "type": "string",
                    "description": "Describe el sexo de los Alumnos"
                },
                "grado": {
                    "type": "number",
                    "description": "Describe el grado de los Alumnos"
                },
                "grupo": {
                    "type": "string",
                    "description": "Describe el grupo de los Alumnos"
                },
                "materias":{
                    "type": "string",
                    "description": "Describe las materias que lleva el Alumno"
                },
                "puntaje":{
                    "type": "string",
                    "description": "Describe las calificaciones del Alumno"
                },
                "promedio":{
                    "type": "number",
                    "description": "Describe el promedio del Alumno"
                }
            }
        },
        "Horarios_Profesores": {
            "type": "object",
            "properties": {
                "id_profesor": {
                    "type": "number",
                    "description": "Identificador único de un Profesor de la Carrera de Computación de la UTM"
                },
                "nombre_profesor":{
                    "type": "string",
                    "description": "Describe el Horario del Profesor"
                },
                "grado":{
                    "type": "number",
                    "description": "Describe el grado del salón al que el Profesor enseña"
                },
                "num_salon":{
                    "type": "string",
                    "description": "Describe el nombre del salón para hubicar la clase"
                },
                "hora":{
                    "type": "string",
                    "description": "Describe las horas de clases impartidas"
                }
            }
        },
        "Horarios_Alumnos": {
            "type": "object",
            "properties": {
                "id_alumno": {
                    "type": "number",
                    "description": "Identificador único de un Profesor de la Carrera de Computación de la UTM"
                },
                "semestre":{
                    "type": "number",
                    "description": "Describe el grado al que el Alumno pertenece"
                },
                "materia":{
                    "type": "string",
                    "description": "Describe los nombres de las materias que llevan los Alumnos"
                },
                "hora":{
                    "type": "string",
                    "description": "Describe las horas de clases impartidas"
                },
                "salon":{
                    "type": "string",
                    "description": "Describe el nombre del salón para hubicarlo"
                }
            }
        },
        "Tutorados": {
            "type": "object",
            "properties": {
                "tutor":{
                    "type": "string",
                    "description": "Describe el nombre del Tutor del Tutorado"
                },
                "id_alu": {
                    "type": "number",
                    "description": "Identificador único de un Tutorado de la Carrera de Computación de la UTM"
                },
                "nombre_alumno":{
                    "type": "string",
                    "description": "Describe el nombre del Tutorado"
                }
            }
        },
        "Tutoria": {
            "type": "object",
            "properties": {
                "id_al":{
                    "type": "string",
                    "description": "Describe la matrícula del Atutorado al Tutor"
                },
                "nombre_prof":{
                    "type": "string",
                    "description": "Describe el nombre del Tutor al Atutorado"
                },
                "nombre_alumno":{
                    "type": "string",
                    "description": "Describe el nombre del Atutorado al Tutor"
                },
                "hora_tutoria":{
                    "type": "string",
                    "description": "Describe la hora en que fue la Tutoría"
                },
                "fecha":{
                    "type": "string",
                    "description": "Describe la fecha de la Tutoría"
                },
                "comentarios":{
                    "type": "string",
                    "description": "Describe lo mencionado por el Atutorado en la Tutoría"
                }
            }
        },
        "Tutores": {
            "type": "object",
            "properties": {
                "id_prof": {
                    "type": "number",
                    "description": "Identificador único de un Tutor para el Atutorado de la Carrera de Computación de la UTM"
                },
                "nombre_profesor":{
                    "type": "string",
                    "description": "Describe el nombre del Tuto al Atutorado"
                },
                "horario_asesoria":{
                    "type": "string",
                    "description": "Describe las horas libres del Tutor para el Atutorado"
                },
                "tutorados":{
                    "type": "string",
                    "description": "Describe los Tutorados del Tutor"
                }
            }
        }
    }
}